refactor(vm): flatten loaders/wazero -> loaders (one engine)
The loader is engine-agnostic and there's only one engine now; drop the
wazero-named subdir. No more "wazero" anywhere in the tree.
