+++ pkg/vm-low-orbit/tests/harness.go
@@ -19,7 +19,7 @@ import (
 	plugins "github.com/taubyte/tau/pkg/vm-low-orbit"
 	"github.com/taubyte/tau/pkg/vm/backend/file"
 	vmContext "github.com/taubyte/tau/pkg/vm/context"
-	loader "github.com/taubyte/tau/pkg/vm/loaders/wazero"
+	loader "github.com/taubyte/tau/pkg/vm/loaders"
 	fileRes "github.com/taubyte/tau/pkg/vm/resolvers/file"
 	source "github.com/taubyte/tau/pkg/vm/sources/taubyte"
 )
