+++ services/substrate/vm.go
@@ -6,7 +6,7 @@ import (
 	dfs "github.com/taubyte/tau/pkg/vm/backend/dfs"
 	"github.com/taubyte/tau/pkg/vm/backend/file"
 	httpBe "github.com/taubyte/tau/pkg/vm/backend/url"
-	loader "github.com/taubyte/tau/pkg/vm/loaders/wazero"
+	loader "github.com/taubyte/tau/pkg/vm/loaders"
 	resolver "github.com/taubyte/tau/pkg/vm/resolvers/taubyte"
 	source "github.com/taubyte/tau/pkg/vm/sources/taubyte"
 )
